Split testing.py into multiple files.
